move loader files
