+++ src/Configuration/Loader/JsonConfigLoader.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Configuration;
+namespace ArtARTs36\MergeRequestLinter\Configuration\Loader;
 
 use ArtARTs36\FileSystem\Contracts\FileSystem;
 use ArtARTs36\MergeRequestLinter\Ci\Credentials\Token;
+use ArtARTs36\MergeRequestLinter\Configuration\Config;
 use ArtARTs36\MergeRequestLinter\Contracts\ConfigLoader;
 use ArtARTs36\MergeRequestLinter\Contracts\ConfigValueTransformer;
 use ArtARTs36\MergeRequestLinter\Contracts\RemoteCredentials;
